// ui/adminv2/apps/web/types/RoutePlugin.ts
/*
 * Copyright (c) 2026. Mikhail Kulik.
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import {Plugin} from "./Plugin";

export interface RoutePlugin {
  readonly id: string
  readonly created_at: string
  route_id: string
  plugin_id: string
  version_constraint: string
  execution_order: number
  config: string | null
  resolved_plugin_version?: string
  readonly plugin?: Plugin
}